refactor: move getComponentClassName to utilities
