+++ slash/react/src/Form/Text/Agent/Text.tsx
@@ -1,7 +1,6 @@
 import "@axa-fr/design-system-slash-css/dist/Form/Text/Agent/Text.agent.scss";
 import { ComponentPropsWithRef, forwardRef } from "react";
-
-import { getComponentClassName } from "../../core";
+import { getComponentClassName } from "../../../utilities";
 
 type Props = Omit<ComponentPropsWithRef<"input">, "required"> & {
   classModifier?: string;
