+++ slash/react/src/Form/Text/Client/Text.tsx
@@ -1,7 +1,7 @@
 import "@axa-fr/design-system-slash-css/dist/Form/Text/Client/Text.client.scss";
 import { ComponentPropsWithRef, forwardRef } from "react";
 
-import { getComponentClassName } from "../../core";
+import { getComponentClassName } from "../../../utilities";
 import { InputError } from "../../InputError";
 
 type Props = Omit<ComponentPropsWithRef<"input">, "required"> & {
