+++ slash/react/src/Layout/Header/Agent/User/User.tsx
@@ -1,7 +1,7 @@
 import { MouseEvent, ReactNode } from "react";
 import { InnerUser } from "./InnerUser";
 import "@axa-fr/design-system-slash-css/dist/Layout/Header/Agent/User/User.scss";
-import { getComponentClassName } from "../../../../agent";
+import { getComponentClassName } from "../../../../utilities";
 
 const defaultClassName = "af-info-user";
