+++ slash/react/src/List/ContentItemDuo/ContentItemDuo.tsx
@@ -2,7 +2,7 @@ import checkIcon from "@material-symbols/svg-400/outlined/check.svg";
 import closeIcon from "@material-symbols/svg-400/outlined/close.svg";
 import { useMemo } from "react";
 import { Variants } from "../../Button/Button.client";
-import { getComponentClassName } from "../../Form/core";
+import { getComponentClassName } from "../../utilities";
 import { Button, Svg } from "../../client";
 
 type ContentItemDuoProps = {
