+++ slash/react/src/Table/Th.tsx
@@ -1,5 +1,5 @@
 import { ComponentPropsWithoutRef } from "react";
-import { getComponentClassName } from "../Form/core";
+import { getComponentClassName } from "../utilities";
 
 type Props = ComponentPropsWithoutRef<"th"> & {
   classModifier?: string;
