+++ slash/react/src/Title/Agent/Title.tsx
@@ -1,7 +1,7 @@
 import "@axa-fr/design-system-slash-css/dist/Title/Agent/Title.scss";
 import { ComponentPropsWithRef, PropsWithChildren, forwardRef } from "react";
 
-import { getComponentClassName } from "../../Form/core";
+import { getComponentClassName } from "../../utilities";
 
 type Headings = "h2" | "h3" | "h4";
