+++ slash/react/src/utilities.ts
@@ -1,1 +1,2 @@
+export { getComponentClassName } from "./utilities/helpers/getComponentClassName";
 export { useIsSmallScreen } from "./utilities/hooks/useIsSmallScreen";
