+++ slash/react/src/utilities/helpers/getComponentClassName.ts
